Integrated back docalign dic-based and hunalign. Integrated back bicleaner training. Integrated biroamer. Multiple fixes
