+++ .gitmodules
@@ -34,3 +34,6 @@
 	path = bleualign-cpp
 	url = https://github.com/bitextor/bleualign-cpp
 	ignore = dirty
+[submodule "biroamer"]
+	path = biroamer
+	url = https://github.com/bitextor/biroamer
